refactor(ksp): Generate dependencies at end
